Move createFlatList into redux selectors
